@amrrus - #Integrity - GET request implemented -addUser -addMovie -addRating -getAllUsers -getAllMovies -getAllRatings -mostRating -leastRating

+++ RatingApp/src/main/java/com/rating/app/ratingRepository.java
@@ -2,6 +2,6 @@ package com.rating.app;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ratingRepository extends CrudRepository<Rating, Long> {
+public interface ratingRepository extends CrudRepository<Rating, Integer> {
 
 }
