+++ RatingApp/src/main/java/com/rating/app/userRepository.java
@@ -2,6 +2,6 @@ package com.rating.app;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface userRepository extends CrudRepository<User, Long>{
+public interface userRepository extends CrudRepository<User, Integer>{
 
 }
